+++ src/Schema/Directives/ValidationDirective.php
@@ -5,7 +5,6 @@ namespace Nuwave\Lighthouse\Schema\Directives;
 use Closure;
 use GraphQL\Type\Definition\ResolveInfo;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
-use Nuwave\Lighthouse\Exceptions\ValidationException;
 use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 use Nuwave\Lighthouse\Support\Contracts\FieldMiddleware;
 use Nuwave\Lighthouse\Support\Traits\HasResolverArguments;
@@ -48,12 +47,18 @@ abstract class ValidationDirective extends BaseDirective implements FieldMiddlew
                         ->make(
                             $args,
                             $this->getRules(),
-                            $this->getMessages()
+                            $this->getMessages(),
+                            // The presence of those custom attributes ensures we get a GraphQLValidator
+                            [
+                                'root' => $root,
+                                'context' => $context,
+                                'resolveInfo' => $resolveInfo,
+                            ]
                         );
 
-                    if ($validator->fails()) {
-                        throw new ValidationException($validator);
-                    }
+                    // The validation exception is caught in the FieldFactory and merged with
+                    // argument directive based validation
+                    $validator->validate();
 
                     return $resolver($root, $args, $context, $resolveInfo);
                 }